fix segundo sumando parseint

+++ suma.js
@@ -1,5 +1,5 @@
 const sum = (a, b) => {
-  return parseInt(a) + parseFloat(b);
+  return parseInt(a) + parseInt(b);
 };
 
 module.exports = sum;
